refactor: finish home hero
